pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "AI-powered job application agent"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "playwright>=1.48.0",
    "sqlmodel>=0.0.21",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "typer>=0.13.0",
    "rich>=13.9.0",
    "feedparser>=6.0.11",
    "google-api-python-client>=2.150.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "apscheduler>=3.10.4",
    "python-docx>=1.2.0",
]

[project.scripts]
job-agent = "job_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_agent"]

[tool.ruff]
src = ["src"]
line-length = 100
